+++ lib_xtcp/src/xc2compat.h
@@ -1,4 +1,5 @@
-// Copyright (c) 2015-2016, XMOS Ltd, All rights reserved
+// Copyright 2015-2025 XMOS LIMITED.
+// This Software is subject to the terms of the XMOS Public Licence: Version 1.
 #ifndef _xc2compat_h_
 #define _xc2compat_h_
 
@@ -14,4 +15,30 @@
 #define alias
 #endif
 
+#ifndef NULLABLE_CLIENT_INTERFACE
+#ifdef __XC__
+#define NULLABLE_CLIENT_INTERFACE(type, name) client interface type ?name
+#else
+#define NULLABLE_CLIENT_INTERFACE(type, name) unsigned name
 #endif
+#endif
+
+#ifdef __XC__
+#define SERVER_INTERFACE_ARRAY(type, name, size) server interface type name[size]
+#else
+#define SERVER_INTERFACE_ARRAY(type, name, size) unsigned *name
+#endif
+
+#ifdef __XC__
+#define CONST_NULLABLE_ARRAY_OF_SIZE(type, name, size) const type (&?name)[size]
+#else
+#define CONST_NULLABLE_ARRAY_OF_SIZE(type, name, size) const type *name
+#endif
+
+#ifdef __XC__
+#define static_const_unsigned static const unsigned
+#else
+#define static_const_unsigned const unsigned
+#endif
+
+#endif /* _xc2compat_h_ */