+++ lib_xtcp/src/xtcp_lwip/xcore/include/arch/cc.h
@@ -60,7 +60,7 @@ typedef u32_t mem_ptr_t;
 #define X8_F "x"
 
 /* Compiler hints for packing structures */
-#ifdef __XC__
+#if __xcore__
 #define PACK_STRUCT_FIELD(x) x
 #else
 #define PACK_STRUCT_FIELD(x) x __attribute__((packed))
@@ -89,7 +89,7 @@ typedef u32_t mem_ptr_t;
 #define LWIP_PLATFORM_HTONL(x) __builtin_bswap32(x)
 
 #include <random.h>
-static random_generator_t rng;
+extern random_generator_t rng;
 
 #define LWIP_RAND() random_get_random_number(&rng)