+++ lib_xtcp/src/xtcp_uip/xtcp_stack.h
@@ -1,2 +1,0 @@
-// Copyright (c) 2017, XMOS Ltd, All rights reserved
-#define XTCP_STACK_UIP 1