+++ lib_xtcp/src/xtcp_uip_includes.h
@@ -1,4 +1,5 @@
-// Copyright (c) 2016-2017, XMOS Ltd, All rights reserved
+// Copyright 2016-2025 XMOS LIMITED.
+// This Software is subject to the terms of the XMOS Public Licence: Version 1.
 #ifndef __xtcp_uip_includes_h__
 #define __xtcp_uip_includes_h__