+++ CHANGELOG.rst
@@ -1,5 +1,14 @@
-TCP/IP library change log
-=========================
+lib_xtcp change log
+===================
+
+UNRELEASED
+----------
+
+  * ADDED: CMake support
+  * FIXED: Many compiler warnings.
+  * CHANGE: Removed 'smi_if i_smi' and 'phy_address' parameters from both
+    xtcp_uip and xtcp_lwip functions. The smi_if now only links the smi 
+    component with the PHY driver.
 
 6.1.0
 -----